customer_notification/requirements.txt
boto3==1.34.131
slack_sdk==3.27.1
python-dotenv==1.0.1
jinja2==3.1.3